move dateRange from constants to utils
