+++ lib/consts.js
@@ -10,22 +10,9 @@ const orderOperations = {
 	download: 'DOWNLOAD',
 };
 
-const dateRange = (start, end) => {
-	if (start && end)
-		return {
-			DateRange: {
-				Start: start,
-				End: end,
-			},
-		};
-
-	return {};
-};
-
 module.exports = {
 	name,
 	version,
 	orderOperations,
 	productString: `${name} ${version}`,
-	dateRange,
 };
