+++ lib/predefinedOrders/PTK.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const utils = require('../consts');
+const utils = require('../utils');
 
 module.exports = (start = null, end = null) => ({
 	version: 'h004',
